api_contidel/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-filter==1.1.0
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
dj-rest-auth==5.0.1